Memory Game App
This is a memory game app designed using Swift, working in XCode. The app can be either be tested on a simulated IOS device using XCode, or uploaded to an IOS device and demonstrated that way.

The goal of the game is to match identical cards, where each card is a different stripped flag. The app shuffles the cards, allows users to play the game by taping the screen and keeps track of points when the game is played against a partner.